+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 * All the public API names are now correctly marked `public` in Julia 1.11 and above. ([#83])
 
+### Changed
+
+* The string `repr` of `DatasetVersion` (e.g. `dataset.versions`) is now valid Julia code. ([#84])
+
 ### Fixed
 
 * The `JuliaHub.update_dataset` function now correctly accepts the `license=(:fulltext, ...)` argument. ([#74])
@@ -146,3 +150,4 @@ Initial package release.
 [#58]: https://github.com/JuliaComputing/JuliaHub.jl/issues/58
 [#74]: https://github.com/JuliaComputing/JuliaHub.jl/issues/74
 [#83]: https://github.com/JuliaComputing/JuliaHub.jl/issues/83
+[#84]: https://github.com/JuliaComputing/JuliaHub.jl/issues/84